Move classes for refactorings logging to separate package.
