+++ src/main/java/org/ml_methods_group/refactoring/logging/MoveMethodRefactoringFeatures.java
@@ -14,27 +14,27 @@
  * limitations under the License.
  */
 
-package org.ml_methods_group.refactoring;
+package org.ml_methods_group.refactoring.logging;
 
 import com.sixrr.metrics.metricModel.MetricsRun;
 import org.jetbrains.annotations.NotNull;
-import org.ml_methods_group.algorithm.refactoring.MoveFieldRefactoring;
+import org.ml_methods_group.algorithm.refactoring.MoveMethodRefactoring;
 
 /**
  * Subclass of {@link RefactoringFeatures} that contains features of a
- * {@link MoveFieldRefactoring}.
+ * {@link MoveMethodRefactoring}.
  */
-public class MoveFieldRefactoringFeatures extends RefactoringFeatures {
+public class MoveMethodRefactoringFeatures extends RefactoringFeatures {
     /**
-     * Extracts features from a given {@link MoveFieldRefactoring}.
+     * Extracts features from a given {@link MoveMethodRefactoring}.
      *
-     * @param refactoring a {@link MoveFieldRefactoring} to extract features from.
+     * @param refactoring a {@link MoveMethodRefactoring} to extract features from.
      * @param metricsRun a result of metrics calculations. Some of metrics values calculated for
      *                   objects given refactoring operates on can be used to extract refactoring
      *                   features.
      */
-    public MoveFieldRefactoringFeatures(
-        final @NotNull MoveFieldRefactoring refactoring,
+    public MoveMethodRefactoringFeatures(
+        final @NotNull MoveMethodRefactoring refactoring,
         final @NotNull MetricsRun metricsRun
     ) {
     }
