+++ src/main/java/org/ml_methods_group/ui/ClassRefactoringPanel.java
@@ -25,8 +25,8 @@ import com.sixrr.metrics.metricModel.MetricsRun;
 import org.jetbrains.annotations.NotNull;
 import org.ml_methods_group.algorithm.refactoring.Refactoring;
 import org.ml_methods_group.config.RefactoringPreferencesLog;
-import org.ml_methods_group.refactoring.RefactoringSessionInfo;
-import org.ml_methods_group.refactoring.RefactoringSessionInfoRenderer;
+import org.ml_methods_group.refactoring.logging.RefactoringSessionInfo;
+import org.ml_methods_group.refactoring.logging.RefactoringSessionInfoRenderer;
 import org.ml_methods_group.utils.ArchitectureReloadedBundle;
 import org.ml_methods_group.utils.ExportResultsUtil;
 import org.ml_methods_group.utils.PsiSearchUtil;
